requirements.txt
pandas==1.5.3
PyMuPDF==1.19.6
pdfplumber==0.5.28
sentence-transformers==2.2.0
faiss-cpu==1.7.2
requests==2.28.1
gradio==3.4.4
torch==1.13.1
